Fix issue of not loading the users org on initial log in
The issue was that loading on mount did not work, because the component
(in this case the MainLayout) was mounted before the user was logged.
Once the user is logged in this is not mounted again.

Moving the mounted hook to the drawer which is only visible once the
user is logged in solved the issue.

Some clean up, so that not only the names of the selected/orgs but
rather the objects are stored and reduced to strings when displaying.