Store refactoring to access membership info
This is a large refactoring of how membership information is handled:
To set the public visibility of an installation, but also for other tasks, the app must know which role a use has in a given organization. This info is contained in the Membership resources but not in the Organization resource. The changes here consist of the following:
- Enhance the store data model for the currently authenticated user.
- Add membership information to this data model.
- In MainLayout, use the membership information of the authenticated user from the store instead of locally fetching related organizations.