Use memberships also on the OrgManagement page
More refactoring to unify and simplify handling of organizations. In the app, only those organizations are of relevance where the authenticated user is a member of. Therefore, I attempt to consistently access organizations through the user's membership info. Because the two are modeled as separate resources, updating one does not automatically update the other (even though they are related in the DB data model). Therefore, when creating a new organization, I now update the membership list in the store.